changed float type to tf.float64

+++ flamedisx/utils.py
@@ -9,7 +9,8 @@ import tensorflow as tf
 
 lgamma = tf.math.lgamma
 o = tf.newaxis
-FLOAT_TYPE = tf.float32
+#FLOAT_TYPE = tf.float32
+FLOAT_TYPE = tf.float64
 INT_TYPE = tf.int32
 
 # Extreme mean and standard deviations give numerical errors
